Refactor `bevy_render_2d::mesh_pipeline`
